everything :tada:
make landing real, not an image
style vineblurb
add activeclassname in navbar
