+++ src/components/Landing/Landing.jsx
@@ -1,14 +1,21 @@
 import React from "react"
 import "./Landing.scss"
-import LandingPic from '../../../static/landing.png'
-import { Scroller } from '../../components'
+import Logo from "../../..//static/circle_logo.png"
+import { Scroller } from "../../components"
 
 const Landing = () => (
-    <div className="landing">
-        <img src={LandingPic} alt="Purple Hyndrangea Project" className="picture" />
-        <a href="#blurb">
-            <Scroller />
-        </a>
+  <div className="actual-main-section">
+    <div className="main-landing-container">
+      <h1 className="big-title">The Purple Hydrangea Project</h1>
+      <img src={Logo} alt="PHP Logo" className="php-logo" />
+      <h2 className="main-subtitle">
+        A purple hydrangea symbolizes the desire for a deeper understanding of
+        someone else.
+      </h2>
     </div>
+    <a href="#blurb">
+      <Scroller />
+    </a>
+  </div>
 )
 export default Landing
